move implementation to function
